Add search select switching
