+++ src/components/Contenteditable.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import {onMounted, ref} from "vue";
+import {onMounted, ref, watch} from "vue";
+import {focusOnEditableElement} from "../utils/focus.ts";
 
 const props = withDefaults(defineProps<{
   tag: string
@@ -13,13 +14,30 @@ const emits = defineEmits<{
   (e: 'update:modelValue', value: string): void
 }>()
 
-const editable = ref(null)
+const editable = ref<HTMLElement | null>(null)
+
+watch(() => props.modelValue, value => {
+  // setTimeout(() => {
+  if (!editable.value) return
+
+  if (value !== editable.value.innerText) {
+    editable.value.innerText = value
+
+    focusOnEditableElement(editable.value)
+  }
+  // if (Math.abs(value.length - editable.value.innerText.length) > 1 || value.slice(0, editable.value.innerText.length) !== editable.value.innerText) {
+  //   editable.value.innerText = value
+  // }
+  // })
+})
 
 onMounted(() => {
+  if (!editable.value) return
+
   editable.value.innerText = props.modelValue
 })
 
-async function onInput(event: InputEvent) {
+async function onInput(event: KeyboardEvent) {
   const target = event.currentTarget as HTMLElement;
   if (!target) return
 
@@ -28,7 +46,7 @@ async function onInput(event: InputEvent) {
 </script>
 
 <template>
-  <component :is="tag" ref="editable" contenteditable="true" @input="onInput($event)">{{ modelValue }}</component>
+  <component :is="tag" ref="editable" contenteditable="true" @keydown="onInput($event)"></component>
 </template>
 
 <style scoped>
